Moved internal classes into dedicated package
