+++ src/commonMain/kotlin/com/ashampoo/xmp/internal/XMPErrorConst.kt
@@ -6,7 +6,7 @@
 // NOTICE:  Adobe permits you to use, modify, and distribute this file in accordance with the terms
 // of the Adobe license agreement accompanying it.
 // =================================================================================================
-package com.ashampoo.xmp
+package com.ashampoo.xmp.internal
 
 internal object XMPErrorConst {
