+++ src/commonMain/kotlin/com/ashampoo/xmp/internal/XMPMetaParser.kt
@@ -6,9 +6,11 @@
 // NOTICE:  Adobe permits you to use, modify, and distribute this file in accordance with the terms
 // of the Adobe license agreement accompanying it.
 // =================================================================================================
-package com.ashampoo.xmp
+package com.ashampoo.xmp.internal
 
-import com.ashampoo.xmp.XMPNormalizer.normalize
+import com.ashampoo.xmp.XMPConst
+import com.ashampoo.xmp.XMPMeta
+import com.ashampoo.xmp.internal.XMPNormalizer.normalize
 import com.ashampoo.xmp.options.ParseOptions
 import nl.adaptivity.xmlutil.dom.Element
 import nl.adaptivity.xmlutil.dom.Node
