+++ src/commonMain/kotlin/com/ashampoo/xmp/internal/XMPNode.kt
@@ -6,8 +6,9 @@
 // NOTICE:  Adobe permits you to use, modify, and distribute this file in accordance with the terms
 // of the Adobe license agreement accompanying it.
 // =================================================================================================
-package com.ashampoo.xmp
+package com.ashampoo.xmp.internal
 
+import com.ashampoo.xmp.XMPConst
 import com.ashampoo.xmp.options.PropertyOptions
 
 /**
