+++ src/commonMain/kotlin/com/ashampoo/xmp/internal/XMPRDFWriter.kt
@@ -6,9 +6,14 @@
 // NOTICE:  Adobe permits you to use, modify, and distribute this file in accordance with the terms
 // of the Adobe license agreement accompanying it.
 // =================================================================================================
-package com.ashampoo.xmp
+package com.ashampoo.xmp.internal
 
-import com.ashampoo.xmp.Utils.escapeXML
+import com.ashampoo.xmp.XMPConst
+import com.ashampoo.xmp.XMPException
+import com.ashampoo.xmp.XMPMeta
+import com.ashampoo.xmp.XMPSchemaRegistry
+import com.ashampoo.xmp.XMPVersionInfo
+import com.ashampoo.xmp.internal.Utils.escapeXML
 import com.ashampoo.xmp.options.SerializeOptions
 
 /**
